% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aovText.R
\name{aovText}
\alias{aovText}
\title{Generates strings of text for the use in markdown documents}
\usage{
aovText(x, sep = "; ", dec = ".", digits = c(3, 3), which.coef = 1)
}
\arguments{
\item{x}{Name of the object containing the result of the analysis from which stats should be extracted.}

\item{sep}{How to separate pieces of information displayed. Defaults to a semicolon followed by a space (sep = "; ").}

\item{dec}{Which decimal separator should be used? Defaults to ".". Allows to quickly changing to a comma in case you are producing a manuscript in German or Portuguese.}

\item{digits}{How many digits should be kept for each piece of numeric information? Defaults to c(3, 3). Importantly, when a P-value rounded to the number of digits hereby defined results in zero, the result is replaced by a more theoretically meaningful alternative (e.g. P < 0.001).}

\item{which.coefs}{Vector indexing which of the potential many variables should be accessed when type = "coefs".}
}
\value{
A string of text to be included in a markdown object.
}
\description{
This function takes as input one object containing a test result and returns its most important informations. Function aovText takes the result of an Anova (returned from a call to the aov function), returning its F (with corresponding model and residuals degrees of freedom) and P-values for a selected variable in the result. In addition, it offers options to customize the output, changing separators and decimal markers (helpful in case you are producing a manuscript in German or Portuguese), and choosing the number of digits to round number to.
}
\examples{
## loading the anorexia dataset
utils::data(anorexia, package = "MASS")
anorex.1 <- aov(Postwt ~ Treat + Prewt, data = anorexia)
aovText(anorex.1, which.coef="Treat")
aovText(anorex.1, which.coef="Prewt")
}
