% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corText.R
\name{corText}
\alias{corText}
\title{Generates strings of text for the use in markdown documents}
\usage{
corText(x, sep = "; ", dec = "default", digits = c(3, 3, 3))
}
\arguments{
\item{x}{Name of the object containing the result of the analysis from which stats should be extracted.}

\item{sep}{How to separate pieces of information displayed. Defaults to a semicolon followed by a space (sep = "; ").}

\item{dec}{Which decimal separator should be used? Defaults to ".". Allows to quickly changing to a comma in case you are producing a manuscript in German or Portuguese.}

\item{digits}{How many digits should be kept for each piece of numeric information? Defaults to c(3, 3, 3). Importantly, when a P-value rounded to the number of digits hereby defined results in zero, the result is replaced by a more theoretically meaningful alternative (e.g. P < 0.001).}
}
\value{
A string of text to be included in a markdown object.
}
\description{
This function takes as input one object containing the result of a correlation test (from a call to the cor.test function) and returns its most important informations. The exact information returned depends on the method used to for the correlation: Pearson's correlation returns r, t and degrees of freedom, and P-value; Spearman returns rho, S, and P-value; and Kendall returns tau, z, and P-value. Function corText offers options to customize the output, like changing separators and decimal markers (helpful in case you are producing a manuscript in German or Portuguese), and choosing the number of digits to round numbers.
}
\examples{
res <- cor.test(~Postwt + Prewt, data=anorexia)
corText(res)
res <- cor.test(~Postwt + Prewt, data=anorexia, method="spearman")
corText(res)
res <- cor.test(~Postwt + Prewt, data=anorexia, method="kendall")
corText(res)
}
