% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genVals.R
\name{getVals}
\alias{getVals}
\title{Generates strings of text for the use in markdown documents}
\usage{
getVals(x, chunk.names = c("statistic", "p.value"))
}
\description{
Generates strings of text for the use in markdown documents
}
