% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.text.R
\name{lm.text}
\alias{lm.text}
\title{Generates strings of text for the use in markdown documents}
\usage{
lm.text(x, sep = "; ", dec = "default", digits = c(3, 3, 3))
}
\arguments{
\item{x}{Name of the object containing the result of the analysis from which stats should be extracted}

\item{sep}{How to separate pieces of information displayed. Defaults to a semicolon followed by a space (sep = "; ").}

\item{dec}{Which decimal separator should be used? Defaults to ".". Allows to quickly changing to a comma in case you are producing a manuscript in German or Portuguese.}

\item{digits}{How many digits should be kept for each piece of numeric information? Defaults to c(3, 3, 3). Importantly, when a P-value rounded to the number of digits hereby defined results in zero, the result is replaced by a more theoretically meaningful alternative (e.g. P < 0.001).}
}
\value{
A string of text to be included in a markdown object.
}
\description{
This function takes as input one object containing a test result and returns its most important informations. Function lm.text takes the result of a regression (returned from a call to the lm function), returning its R², F-value (with corresponding model and residuals degrees of freedom) and P-value. In addition, they offer options to customize the output, like replacing names of some parameters, changing separators and decimal markers (helpful in case you are producing a manuscript in German or Portuguese), and choosing the number of digits to round number to.
}
\examples{
## loading the anorexia dataset
utils::data(anorexia, package = "MASS")
anorex.1 <- lm(Postwt ~ Prewt, data = anorexia)
lm.text(anorex.1)
}
