% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.sd.text.R
\name{mean.sd.text}
\alias{mean.sd.text}
\title{Generates strings of text for the use in markdown documents}
\usage{
\method{mean}{sd.text}(x, dec = "default", digits = c(1, 1))
}
\arguments{
\item{x}{A numeric vector}

\item{dec}{Which decimal separator should be used? Defaults to ".". Allows to quickly changing to a comma in case you are producing a manuscript in German or Portuguese.}

\item{digits}{How many digits should be kept for each piece of numeric information? Defaults to c(1, 1).}
}
\value{
A string of text to be included in a markdown object.
}
\description{
This function takes as input one vector and returns its mean and standard deviation in the most standard notation (mean ± sd).
}
\examples{
x <- rnorm(50, mean=10, sd=20)
mean.sd.text(x)
# Rounding 
mean.sd.text(x, digits=c(1, 1))
}
