% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.list.R
\name{tidy.list}
\alias{tidy.list}
\title{Generates a neat list of elements}
\usage{
tidy.list(x, sep = ",", last = "and", dec = ".")
}
\arguments{
\item{x}{Name of the object containing character vector.}

\item{sep}{How to separate pieces of information displayed. Defaults to a commas followed by a space.}

\item{last}{How to separate the last element. Defaults to "and".}
}
\value{
A string of text to be used in markdown documents.
}
\description{
This function takes as input one character vector and constructs a meaningful text. By default it results in a list with elements separated by commas and the last element separated by ", and".
}
\examples{
x <- c("apples", "oranges", "grapes")
tidy.list(x)
tidy.list(c(2.3, 5.4, 3), dec=",")
}
