% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tText.R
\name{tText}
\alias{tText}
\title{Generates strings of text for the use in markdown documents}
\usage{
tText(x, df = "df", sep = "; ", dec = "default", digits = c(3, 1, 3))
}
\arguments{
\item{x}{Name of the object containing the result of the analysis from which stats should be extracted}

\item{df}{How to word the abbreviation for degrees of freedom. Defaults to "df".}

\item{sep}{How to separate pieces of information displayed. Defaults to a semicolon followed by a space (sep = "; ").}

\item{dec}{Which decimal separator should be used? Defaults to ".". Allows to quickly changing to a comma in case you are producing a manuscript in German or Portuguese.}

\item{digits}{How many digits should be kept for each piece of numeric information? Defaults to c(3, 1, 3). Importantly, when a P-value rounded to the number of digits hereby defined results in zero, the result is replaced by a more theoretically meaningful alternative (e.g. P < 0.001).}
}
\value{
A string of text to be included in a markdown object.
}
\description{
This function takes as input one object containing a test result and returns its most important informations. Function t.text takes the result of a t-test (returned from a call to the t.test function), returning its t-value, number of degrees of freedom (possibly corrected) and the test P-value. It offers options to customize the output, like replacing names of some parameters, changing separators and decimal markers (helpful in case you are producing a manuscript in German or Portuguese), and choosing the number of digits to round number to.
}
\examples{
res <- t.test(extra ~ group, data = sleep)
t.text(res)
# Changing decimal marker in case you are producing a manuscript in German or Portuguese 
t.text(res, dec=",")
}
