% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vals.text.R
\name{valsText}
\alias{valsText}
\title{Rounds and changes decimal markers in single numeric values}
\usage{
valsText(x, dec = "default", digits = 3)
}
\arguments{
\item{x}{Name of the object containing the numeric value.}

\item{dec}{Which decimal separator should be used? Defaults to ".". Allows to quickly changing to a comma in case you are producing a manuscript in German or Portuguese.}

\item{digits}{To how many digitis should the value be rounded to? Default is 3.}
}
\value{
A string of text to be used in markdown documents.
}
\description{
This function takes as input one numeric value, rounds and changes decimal markers whenever necessary.
}
\examples{
x <- 5.46923
vals.text(x, dec=",", digits=2)
}
