% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxText.R
\name{wilcoxText}
\alias{wilcoxText}
\title{Generates strings of text for the use in markdown documents}
\usage{
wilcoxText(x, df = NULL, sep = "; ", dec = "default", digits = c(3, 3))
}
\arguments{
\item{x}{Name of the object containing the result of the analysis from which stats should be extracted.}

\item{df}{Allows to included the number of degrees of freedom there were in the data. This information cannot be retrieved from the object containing the result of a call to wilcox.test. For two-sample cells it seems to be n1+n2, but (n1+n2)/2 for a paired test.}

\item{sep}{How to separate pieces of information displayed. Defaults to a semicolon followed by a space (sep = "; ").}

\item{dec}{Which decimal separator should be used? Defaults to ".". Allows to quickly changing to a comma in case you are producing a manuscript in German or Portuguese.}

\item{digits}{How many digits should be kept for each piece of numeric information? Defaults to c(3, 3). Importantly, when a P-value rounded to the number of digits hereby defined results in zero, the result is replaced by a more theoretically meaningful alternative (e.g. P < 0.001).}
}
\value{
A string of text to be included in a markdown object.
}
\description{
This function takes as input one object containing the result of a one- and two-sample Wilcoxon tests on vectors of data (from a call to the wilcox.test function) and returns its most important informations. The exact information returned depends on the method used for the correlation: Pearson's correlation returns r, t and degrees of freedom, and P-value; Spearman returns rho, S, and P-value; and Kendall returns tau, z, and P-value. Function corText offers options to customize the output, like changing separators and decimal markers (helpful in case you are producing a manuscript in German or Portuguese), and choosing the number of digits to round numbers to.
}
\examples{
res <- wilcox.test(anorexia$Postwt, anorexia$Prewt)
wilcoxText(res, df=length(anorexia$Postwt) + 
  length(anorexia$Prewt))
res <- wilcox.test(anorexia$Postwt, anorexia$Prewt, paired=T)
wilcoxText(res, df=length(anorexia$Postwt))
}
